added icons autocentering in main screen
After not being able to center the list items of the ul.icons
horizontally on screen, i tried with javascript/jquery and was able to
do it.
However i think this is very heavy on the CPU usage, beacuse it makes a
fuck-ton of calculations everytime the window is resized. I have to
further research on another way of doing this via css or smth...